split into components and modules
