+++ package.json
@@ -15,6 +15,7 @@
     "lodash": "^3.10.1",
     "react": "^0.14.2",
     "react-dom": "^0.14.2",
+    "react-redux": "^4.0.0",
     "redux": "^3.0.4"
   },
   "devDependencies": {
