+++ store.js
@@ -1,0 +1,6 @@
+'use strict';
+
+var createStore = require('redux').createStore;
+var rootReducer = require('./reducers/calculate');
+
+module.exports = createStore(rootReducer);
