+++ styles.js
@@ -1,0 +1,23 @@
+'use strict';
+
+module.exports = {
+	app: {
+		paddingLeft: '3rem'
+	},
+	label: {
+		flexBasis: '30%'
+	},
+	field: {
+		display: 'flex',
+		padding: '.5rem'
+	},
+	input: {
+		fontSize: '1rem',
+		padding: '.3rem'
+	},
+	value: {
+		flexBasis: '65%',
+		height: '2rem'
+	}
+};
+
